+++ CHANGELOG.md
@@ -6,4 +6,5 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [0.1.0] Unreleased
-
+### Added
+- Add cmdlet help for Start-DmarcRuaMailboxScrape.